+++ streamlit/component-template/template-vite/vite_vanilla_component/__init__.py
@@ -23,7 +23,7 @@ if not _RELEASE:
         # We give the component a simple, descriptive name ("my_component"
         # does not fit this bill, so please choose something better for your
         # own component :)
-        "my_component",
+        "vite_vanilla_component",
         # Pass `url` here to tell Streamlit that the component will be served
         # by the local dev server that you run via `npm run start`.
         # (This is useful while your component is in development.)
@@ -34,8 +34,8 @@ else:
     # replace the `url` param with `path`, and point it to to the component's
     # build directory:
     parent_dir = os.path.dirname(os.path.abspath(__file__))
-    build_dir = os.path.join(parent_dir, "frontend/build")
-    _component_func = components.declare_component("my_component", path=build_dir)
+    build_dir = os.path.join(parent_dir, "frontend/dist")
+    _component_func = components.declare_component("vite_vanilla_component", path=build_dir)
 
 
 # Create a wrapper function for the component. This is an optional
@@ -44,7 +44,7 @@ else:
 # our component's API: we can pre-process its input args, post-process its
 # output value, and add a docstring for users.
 def my_component(name, key=None):
-    """Create a new instance of "my_component".
+    """Create a new instance of "vite_vanilla_component".
 
     Parameters
     ----------